pyproject.toml
[tool.poetry]
name = "echis_web2"
version = "0.0.1"
description = ""
authors = ["kengenal"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^1.1.2"
flask-mongoengine = "^1.0.0"
PyJWT = "^1.7.1"
Flask-Session = "^0.3.2"

[tool.poetry.dev-dependencies]
pytest = "^6.1.2"
factory_boy = "3.1.0"
mongomock = "^3.21.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
